Create Exercise2.java
Exercise 2
Write a program to represent a class called Point3D that stores the coordinates of a point in 3D
space. Add a constructor to initialize the variables passed. Add a method called distance () to
calculate the distance from origin to a point.
Create another class called PointApp with the main function and create two objects of the class
Point3D. Calculate and display the distance to the origin of these objects.
Note: distance = sqrt( x
2 +y2 + z2
)